+++ Website/htdocs/mpmanager/app/Http/Controllers/PVController.php
@@ -88,8 +88,8 @@ class PVController extends Controller
 
     private function formatEnergyData($val): float
     {
-        $val = (double)(str_replace('.', '', $val));
-        $val = (double)(str_replace(',', '.', $val));
+        $val = (float)(str_replace('.', '', $val));
+        $val = (float)(str_replace(',', '.', $val));
         return $val;
     }
 }
